requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-property==0.2.2
asyncpg==0.30.0
attrs==25.1.0
cached-property==1.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
deprecation==2.1.0
ecdsa==0.19.0
fastapi==0.115.7
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
jwcrypto==1.5.6
jwt==1.3.1
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
python-jose==3.3.0
python-keycloak==5.1.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
